backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pymupdf>=1.24.11
pydantic==2.8.2
python-dotenv==1.0.1
reportlab==4.2.2
httpx==0.27.2
